+++ sink_appendtofile.go
@@ -52,6 +52,10 @@ func AppendToFile(filename string) Command {
 		// expand our input
 		expFilename := p.Env.Expand(filename)
 
+		// debugging support
+		Tracef("AppendToFile(%#v)", filename)
+		Tracef("=> AppendToFile(%#v)", expFilename)
+
 		// open / create the file
 		fh, err := os.OpenFile(expFilename, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
 		if err != nil {
@@ -66,6 +70,7 @@ func AppendToFile(filename string) Command {
 		if p.Flags&contextIsPipeline != 0 {
 			// we are part of a pipe
 			for line := range p.Stdin.ReadLines() {
+				TraceOutput("file", "%s", line)
 				_, err = fh.WriteString(line)
 				if err != nil {
 					return StatusNotOkay, err
@@ -78,6 +83,7 @@ func AppendToFile(filename string) Command {
 		} else {
 			// we are part of a list
 			for line := range p.Stdout.ReadLines() {
+				TraceOutput("file", "%s", line)
 				_, err = fh.WriteString(line)
 				if err != nil {
 					return StatusNotOkay, err