+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   - Lists now write status code and error message when tracing enabled
   - Pipelines now write status code and error message when tracing enabled
   - `And()` now supports tracing
+  - `AppendToFile()` now supports tracing
   - `AppendToTempFile()` now supports tracing
   - `Chmod()` now supports tracing
   - `CountLines()` now supports tracing